requirements.txt
django >= 3.0
django-markup
Pillow
pygments
django-filter
django-bleach
httpie
django-oauth-toolkit
django-cors-middleware
coverage
cryptography
djangorestframework >=3.10.3
django-cors-headers >=3.1.1
drf-jwt >=1.13.3
pyjwt 
requests
djongo
sqlparse
dnspython
python-jose 
social-auth-app-django >= 3.0.0
django-rest-knox
Psycopg2
fontawesome-free == 5.15.3
django-bootstrap-v5
